requirements.txt
tornado>=6.2
requests
pyTelegramBotAPI

python-telegram-bot==12.7
urllib3
matplotlib
pickle-mixin
telepot
pytz
tzwhere
bs4
